+++ DEnc/Commands/Mp4BoxCommandBuilder.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using DEnc.Models;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
