+++ DEnc/Constants.cs
@@ -1,4 +1,5 @@
 ﻿using DEnc.Commands;
+using DEnc.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
