+++ DEnc/Encode/QualityCrusher.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using DEnc.Models;
+using DEnc.Models.Interfaces;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
