+++ DEncTests/ConfigTests.cs
@@ -1,4 +1,5 @@
 ﻿using DEnc;
+using DEnc.Models;
 using System;
 using System.Collections.Generic;
 using System.IO;
