+++ DEncTests/EncodingTests.cs
@@ -1,4 +1,5 @@
 ﻿using DEnc;
+using DEnc.Models;
 using NaiveProgress;
 using System;
 using System.Collections.Generic;
